refactor(react): 拆出 TeamBridge 模块，ReactEngine 剥离多代理职责
把 ReactEngine 中所有与 TeamContext 相关的逻辑收敛到独立的
react/team_bridge.rs 模块，使主 ReAct 状态机不再直接承担多代理调度职责。

迁移到 team_bridge.rs：
- SubAgentDrainResult / SubResult / SubAgentFuture / ActiveSubAgent 类型
- sub_agent_stream_message / send_sub_agent_output / spawn_sub_agent_stream_forwarder
  子 Agent 流式事件转发（独立线程，把子 Agent StreamEvent 翻译为父级 AgentOutput）
- format_team_roster 团队花名册渲染
- spawn_ready_sub_agents / drain_sub_agent_inboxes / drain_main_agent_messages
  （第二个 impl ReactEngine 块整体迁移，跨文件 impl 仍挂在同一类型上）

ReactEngine 字段可见性调整为 pub(super)，供同级 team_bridge 模块访问；
engine.rs 顺带清理不再使用的 Future/Pin/FuturesUnordered/StreamExt 导入。

效果：engine.rs 从 2957 行降至 2129 行，多代理逻辑与单 Agent ReAct 循环解耦，
为后续按事件接口进一步解耦 TeamContext 铺路。行为无变化，208 测试全通过。
