+++ crates/tiangong-core/src/react/mod.rs
@@ -1,3 +1,4 @@
 pub mod context;
 pub mod engine;
 pub mod message;
+mod team_bridge;
